refactor(web): single-source isFixResolvedClean in the drift hook
